moved spiff conversion functions to FormattingService and use it in InstructionsForEndUser w/ burnettk
